versioing

+++ backend/GenerateAppWPF/App.xaml.cs
@@ -18,7 +18,28 @@ namespace GenerateAppWPF
     /// </summary>
     public partial class App : Application
     {
-        private Process process;
+        private static Process process;
+        public static void StartApp()
+        {
+            try
+            {
+                if (process != null && !process.HasExited)
+                {
+                    process.Kill(true);
+                }
+                var dir = Path.Combine(Environment.CurrentDirectory, "publish");
+                var psi = new ProcessStartInfo(Path.Combine(dir, "GenerateFromDb.exe"));
+                psi.WorkingDirectory = dir;
+                psi.CreateNoWindow = true;
+                process = Process.Start(psi);
+
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex.ToString());
+            }
+
+        }
         public App()
         {
             Console.WriteLine("start");
@@ -29,33 +50,15 @@ namespace GenerateAppWPF
 
         private void App_Exit(object sender, ExitEventArgs e)
         {
-            if (!process.HasExited)
+            if (process != null  && !process.HasExited)
                 process.Kill(true);
         }
 
         private void App_Startup(object sender, StartupEventArgs e)
         {
 
-            Task.Run(() =>
-            // GenerateFromDB.Program.Main(new string[1] { "--urls http://localhost:9100" })
-            {
-              try
-              {
-                var dir = Path.Combine(Environment.CurrentDirectory, "publish");
-                var psi = new ProcessStartInfo(Path.Combine(dir, "GenerateFromDb.exe"));
-                psi.WorkingDirectory = dir;
-                psi.CreateNoWindow = true;
-                process = Process.Start(psi);
-
-                
-              }              
-                 catch (Exception ex)
-                 {
-                     Console.WriteLine(ex.ToString());
-                 }
-             }
-            );  
-            //GenerateFromDB.Program.Main(e.Args);
+            Task.Run(() =>{ StartApp();});  
+            
         }
     }
 }
